fix: response and null checker

+++ src/modules/task/task.service.ts
@@ -42,35 +42,64 @@ export async function getSingleTask(userId: string, taskId: string) {
 }
 
 export async function createTask(userId: string, task: createTaskType) {
+    const { tags, ...rest } = task;
     const newTask = await prisma.task.create({
         data: {
             userId,
-            ...task,
+            TaskTag: {
+                connect: tags?.map(tag => ({
+                    name: tag
+                }))
+            },
+            ...rest
         }
+    }).catch((error) => {
+        console.log(error);
+        throw new AppError("Failed to create task", 500);
     });
+        
 
     return newTask;
 }
 
 export async function createTasks(userId: string, body: createBulkTasksType) {
-    const newTasks = await prisma.task.createMany({
-        data: body.tasks.map(task => ({
-            userId,
-            ...task
-        }))
+    const newTasks = await Promise.all(
+        body.tasks.map(async (task) => {
+            const { tags, ...rest } = task;
+            return await prisma.task.create({
+                data: {
+                    userId,
+                    ...rest,
+                    TaskTag: {
+                        connect: tags?.map(tag => ({
+                            name: tag
+                        }))
+                    }
+                }
+            });
+        })
+    ).catch((error) => {
+        console.log(error);
+        throw new AppError("Failed to create tasks", 500);
     });
 
     return newTasks;
 }
 
 export async function updateTask(userId: string, taskId: string, task: updateTaskType) {
+    const { tags, ...rest } = task;
     const updatedTask = await prisma.task.update({
         where: {
             id: taskId,
             userId
         },
         data: {
-            ...task
+            ...rest,
+            TaskTag: {
+                connect: tags?.map(tag => ({
+                    name: tag
+                }))
+            }
         }
     });
 
@@ -82,6 +111,17 @@ export async function updateTask(userId: string, taskId: string, task: updateTas
 }
 
 export async function deleteTask(userId: string, taskId: string) {
+    const task = await prisma.task.findFirst({
+        where: {
+            id: taskId,
+            userId
+        }
+    });
+
+    if (!task) {
+        throw new AppError("Task not found", 404);
+    }
+
     const deletedTask = await prisma.task.delete({
         where: {
             id: taskId,
@@ -89,14 +129,23 @@ export async function deleteTask(userId: string, taskId: string) {
         }
     });
 
-    if (!deletedTask) {
-        throw new AppError("Task not found", 404);
-    }
-
     return deletedTask;
 }
 
 export async function deleteTasks(userId: string, body: deleteTasksType) {
+    const tasks = await prisma.task.findMany({
+        where: {
+            id: {
+                in: body.tasks
+            },
+            userId
+        }
+    });
+
+    if (tasks.length === 0) {
+        throw new AppError("No tasks found to delete", 404);
+    }
+
     const deletedTasks = await prisma.task.deleteMany({
         where: {
             id: {
